+++ packages/core/src/worldgen/index.ts
@@ -1,0 +1,6 @@
+export { deriveWorldType } from "./deriveWorldType.js";
+export { generateWorld } from "./generateWorld.js";
+export { validateWorld } from "./validateWorld.js";
+export { attemptSeed } from "./attemptSeed.js";
+export type { WorldTypeId, WorldConfig, WorldGenResult } from "./types.js";
+export { getWorldConfig } from "./worldConfigs.js";